ai_doc_nal/requirements.txt
langchain==0.0.128
llama-index==0.5.4
Pillow==9.5.0
streamlit==1.19.0
PyPDF2==3.0.1
